Final changes for all SqlServer classes - SqlServer is now fully supported. Also some updates where Int32 was being used (to Int64). Tested on sqlite3 & sqlserver. For now sqlserver is inline sql but later will be converted to StoredProcs.

+++ LibreStore/Models/DbCommon.cs
@@ -31,7 +31,7 @@ public class DbCommon : DbCommonConnection, IDbCommon{
         return mainTokenId;
     }
 
-    public Int32 DeleteBucket(DbConnection Connection, DbCommand Command){
+    public Int64 DeleteBucket(DbConnection Connection, DbCommand Command){
         try{
             Console.WriteLine("DeleteBucket...");
             Connection.Open();
@@ -51,7 +51,7 @@ public class DbCommon : DbCommonConnection, IDbCommon{
         }
     }
 
-    public int GetOrInsert(){
+    public Int64 GetOrInsert(){
         try{
             Console.WriteLine("GetOrInsert...");
             DbConnection.Open();
@@ -59,7 +59,7 @@ public class DbCommon : DbCommonConnection, IDbCommon{
             using (var reader = DbCommand.ExecuteReader())
             {
                 reader.Read();
-                var id = reader.GetInt32(0);
+                var id = reader.GetInt64(0);
                 Console.WriteLine($"GetOrInsert() id: {id}");
                 reader.Close();
                 return id;
